feat: 시도 추출
