+++ fronend/pages/home/stay.tsx
@@ -5,16 +5,5 @@ import {View} from 'react-native';
 import ReligionSelects from '../components/religionSelects';
 
 export const StayPage = () => {
-  useEffect(() => {
-    const fetchData = async () => {
-      const sido = await getSidoCode().then(
-        res => res.response.result.featureCollection.features,
-      );
-      console.log(sido);
-    };
-
-    fetchData();
-  }, []);
-
   return <ReligionSelects />;
 };
